feat : 간단히 보기 차기 만들기
